requirements.txt
discord.py>=2.5.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
aiosqlite>=0.20.0,<1.0.0
Flask>=3.0.0,<4.0.0